Just wanted to update my links so it's just 1
I added new charts to both the 2017 and 2018 folders. I think I fixed a couple things in the old ones as well (but I'm not sure if I uploaded the updates...)
I'm not sure how many of the 2017/2018 folders are on chorus. If someone could go through and add them to chorus for me if they're not found it would be appreciated a ton. Also if I rejoin the discord, please let me be a charter. I've made over 100 over the last 2 years (and I got way better at it) and I'm tired of being shut down every time I ask about it...